# server/swagger.yaml
openapi: 3.0.0
info:
  title: ZUCT
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /exams/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - exams
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - exams
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exams"
    post:
      summary: CRUD ACTION update
      tags:
        - exams
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exams"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/exams"
  /exams/findBy_courses/{key}:
    get:
      summary: CRUD ACTION findBy_courses
      tags:
        - exams
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exams"
  /exams/findBy_lecturer/{key}:
    get:
      summary: CRUD ACTION findBy_lecturer
      tags:
        - exams
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exams"
  /exams/findBy_students/{key}:
    get:
      summary: CRUD ACTION findBy_students
      tags:
        - exams
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exams"
  /exams/{id}/validate:
    post:
      summary: This API is used to validate the exam
      security: []   # No security
      tags:
        - exams
      parameters:
        - name: id
          in: path
          description: is of the exam
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  lecturers  -----
...
  /lecturers/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - lecturers
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - lecturers
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/lecturers"
    post:
      summary: CRUD ACTION update
      tags:
        - lecturers
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/lecturers"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/lecturers"
  /lecturers/findBy_courses/{key}:
    get:
      summary: CRUD ACTION findBy_courses
      tags:
        - lecturers
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/lecturers"
    
  #-----  students  -----
...
  /students/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - students
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - students
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/students"
    post:
      summary: CRUD ACTION update
      tags:
        - students
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/students"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/students"
  /students/findBy_courses/{key}:
    get:
      summary: CRUD ACTION findBy_courses
      tags:
        - students
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/students"
    
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    exams:
      properties:
        _id:
          type: string
        place:
          type: string
        score:
          type: number
        valid:
          type: boolean
      required:
        -  _id
    lecturers:
      properties:
        _id:
          type: string
        first_name:
          type: string
        surname:
          type: string
      required:
        -  _id
    students:
      properties:
        _id:
          type: string
        address:
          type: string
        dob:
          type: integer
        first_name:
          type: string
        phone:
          type: number
        surname:
          type: string
      required:
        -  _id
        - first_name